+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"29.05.2026 08:37"
-#define NE_VERSION   L"2026.05.29.08"
+#define NE_PUBLISHED L"29.05.2026 09:57"
+#define NE_VERSION   L"2026.05.29.09"